Ignore case in checking guess
It has always felt like a bug to me that guesses with mismatched case
would be incorrect and I've watched players get frustrated with this
issue many times.

**Possible issues**: languages with semantic case:

For example in German the case of the first letter determines if the
word is a noun and in some cases can change the meaning of the word
(example needed)

However, this is an edge case as even in these languages it is rare that
the case changes the meaning. Furthermore, the only languages currently
supported officially are English and Japanese. It seems reasonable to me
to make this patch and revisit if necessary.